blue spell.requirements.txt
streamlit
pyspellchecker
nltk
pandas